Migrate to rekt cronjob package (#6800)
Use the cronjob package from reconciler-test
