--- test/rekt/resources/job/job.yaml
@@ -1,41 +1,0 @@
-# Copyright 2022 The Knative Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: batch/v1
-kind: CronJob
-metadata:
-  name: {{ .name }}
-  namespace: {{ .namespace }}
-spec:
-  schedule: "* * * * *"
-  jobTemplate:
-    metadata:
-      labels:
-        app: {{ .name }}
-    spec:
-      template:
-        spec:
-          containers:
-          - name: user-container
-            image: {{ .image }}
-            imagePullPolicy: IfNotPresent
-            env:
-            - name: ONE_SHOT
-              value: "true"
-            - name: POD_NAME
-              value: heartbeats
-            - name: POD_NAMESPACE
-              value: {{ .namespace }}
-          restartPolicy: Never
-
